Move App files to root and update imports
Renamed and moved App.jsx and App.css from src/App/ to the src/ root directory. Updated import paths in App.jsx and main.jsx to reflect the new file locations.
