+++ src/App.jsx
@@ -1,7 +1,7 @@
 import AuthProvider from '@lib/auth/authProvider/AuthProvider';
 import { BrowserRouter } from 'react-router-dom'
 import Router from '@routes/root';
-import '@/App/App.css';
+import './App.css';
 
 function App() {
   return (
