+++ src/main.jsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom/client'
-import App from './App/App'
+import App from './App'
 import React from 'react'
 import './index.css'
